ci: Add job to run linters
This adds the configuration to install and run golangci-lint to manage
the execution of multiple linters.

To install the linters run:

    make golangci-lint

Or to install it along with other tools:

    make tools

Then to run the linters use:

    make lint

The linters can be configured to run in an editor as well.

See: https://golangci-lint.run/usage/integrations/#editor-integration